+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { ProposalDataListItemSkeleton, type IProposalDataListItemSkeletonProps } from './proposalDataListItemSkeleton';